+++ resources/views/livewire/rack-katalog-vorschau.blade.php
@@ -1,0 +1,4 @@
+{{-- Ein Wurzelelement, wie Livewire es verlangt. --}}
+<div>
+    <x-rack.vorschau :appearance="$appearance" :he="$he" :image="$bild" />
+</div>